Fully mitigates #46
