+++ include/boost/numeric/ublas/tensor/expression_operator.hpp
@@ -6,7 +6,7 @@
 //  http://www.boost.org/LICENSE_1_0.txt)
 //
 //  The authors gratefully acknowledge the support of
-//  Fraunhofer and Google in producing this work
+//  and Google in producing this work
 //  which started as a Google Summer of Code project.
 
 #ifndef BOOST_UBLAS_EXPRESSION_OPERATOR_HPP
@@ -14,6 +14,8 @@
 
 #include "multi_index_utility.hpp"
 #include "tensor_cast_macros.hpp"
+#include "functions.hpp"
+#include "expression_relational_operator.hpp"
 #include <boost/yap/user_macros.hpp>
 
 namespace boost::numeric::ublas {
@@ -53,30 +55,6 @@ BOOST_YAP_USER_UDT_ANY_BINARY_OPERATOR(
     divides, boost::numeric::ublas::detail::tensor_expression,
     boost::numeric::ublas::is_tensor)
 
-BOOST_YAP_USER_UDT_ANY_BINARY_OPERATOR(
-    less, boost::numeric::ublas::detail::tensor_expression,
-    boost::numeric::ublas::is_tensor)
-
-BOOST_YAP_USER_UDT_ANY_BINARY_OPERATOR(
-    less_equal, boost::numeric::ublas::detail::tensor_expression,
-    boost::numeric::ublas::is_tensor)
-
-BOOST_YAP_USER_UDT_ANY_BINARY_OPERATOR(
-    equal_to, boost::numeric::ublas::detail::tensor_expression,
-    boost::numeric::ublas::is_tensor)
-
-BOOST_YAP_USER_UDT_ANY_BINARY_OPERATOR(
-    not_equal_to, boost::numeric::ublas::detail::tensor_expression,
-    boost::numeric::ublas::is_tensor)
-
-BOOST_YAP_USER_UDT_ANY_BINARY_OPERATOR(
-    greater, boost::numeric::ublas::detail::tensor_expression,
-    boost::numeric::ublas::is_tensor)
-
-BOOST_YAP_USER_UDT_ANY_BINARY_OPERATOR(
-    greater_equal, boost::numeric::ublas::detail::tensor_expression,
-    boost::numeric::ublas::is_tensor)
-
 // Expr to Expr
 BOOST_YAP_USER_BINARY_OPERATOR(plus,
                                boost::numeric::ublas::detail::tensor_expression,
@@ -96,24 +74,7 @@ BOOST_YAP_USER_UNARY_OPERATOR(unary_plus,
 BOOST_YAP_USER_UNARY_OPERATOR(negate,
                               boost::numeric::ublas::detail::tensor_expression,
                               boost::numeric::ublas::detail::tensor_expression)
-BOOST_YAP_USER_BINARY_OPERATOR(equal_to,
-                               boost::numeric::ublas::detail::tensor_expression,
-                               boost::numeric::ublas::detail::tensor_expression)
-BOOST_YAP_USER_BINARY_OPERATOR(less,
-                               boost::numeric::ublas::detail::tensor_expression,
-                               boost::numeric::ublas::detail::tensor_expression)
-BOOST_YAP_USER_BINARY_OPERATOR(not_equal_to,
-                               boost::numeric::ublas::detail::tensor_expression,
-                               boost::numeric::ublas::detail::tensor_expression)
-BOOST_YAP_USER_BINARY_OPERATOR(greater,
-                               boost::numeric::ublas::detail::tensor_expression,
-                               boost::numeric::ublas::detail::tensor_expression)
-BOOST_YAP_USER_BINARY_OPERATOR(greater_equal,
-                               boost::numeric::ublas::detail::tensor_expression,
-                               boost::numeric::ublas::detail::tensor_expression)
-BOOST_YAP_USER_BINARY_OPERATOR(less_equal,
-                               boost::numeric::ublas::detail::tensor_expression,
-                               boost::numeric::ublas::detail::tensor_expression)
+
 // Tensor Contraction
 template <class tensor_type_left, class tuple_type_left,
           class tensor_type_right, class tuple_type_right>
@@ -150,7 +111,8 @@ auto operator*(std::pair<tensor_type_left const &, tuple_type_left> lhs,
 template <class T, class F, class V, class Expr>
 auto operator+=(boost::numeric::ublas::tensor<T, V, F> &lhs, Expr &&e) {
   decltype(auto) expr =
-      boost::yap::as_expr<boost::numeric::ublas::detail::tensor_expression>(std::forward<Expr>(e));
+      boost::yap::as_expr<boost::numeric::ublas::detail::tensor_expression>(
+          std::forward<Expr>(e));
   auto shape = boost::yap::transform(
       expr, boost::numeric::ublas::detail::transforms::get_extents{});
   if (shape != lhs.extents() && !shape.is_free_scalar()) {
@@ -166,7 +128,8 @@ auto operator+=(boost::numeric::ublas::tensor<T, V, F> &lhs, Expr &&e) {
 template <class T, class F, class V, class Expr>
 auto operator-=(boost::numeric::ublas::tensor<T, V, F> &lhs, Expr &&e) {
   decltype(auto) expr =
-      boost::yap::as_expr<boost::numeric::ublas::detail::tensor_expression>(std::forward<Expr>(e));
+      boost::yap::as_expr<boost::numeric::ublas::detail::tensor_expression>(
+          std::forward<Expr>(e));
   auto shape = boost::yap::transform(
       expr, boost::numeric::ublas::detail::transforms::get_extents{});
   if (shape != lhs.extents() && !shape.is_free_scalar()) {
@@ -181,7 +144,8 @@ auto operator-=(boost::numeric::ublas::tensor<T, V, F> &lhs, Expr &&e) {
 template <class T, class F, class V, class Expr>
 auto operator*=(boost::numeric::ublas::tensor<T, V, F> &lhs, Expr &&e) {
   decltype(auto) expr =
-      boost::yap::as_expr<boost::numeric::ublas::detail::tensor_expression>(std::forward<Expr>(e));
+      boost::yap::as_expr<boost::numeric::ublas::detail::tensor_expression>(
+          std::forward<Expr>(e));
   auto shape = boost::yap::transform(
       expr, boost::numeric::ublas::detail::transforms::get_extents{});
   if (shape != lhs.extents() && !shape.is_free_scalar()) {
@@ -196,7 +160,8 @@ auto operator*=(boost::numeric::ublas::tensor<T, V, F> &lhs, Expr &&e) {
 template <class T, class F, class V, class Expr>
 auto operator/=(boost::numeric::ublas::tensor<T, V, F> &lhs, Expr &&e) {
   decltype(auto) expr =
-      boost::yap::as_expr<boost::numeric::ublas::detail::tensor_expression>(std::forward<Expr>(e));
+      boost::yap::as_expr<boost::numeric::ublas::detail::tensor_expression>(
+          std::forward<Expr>(e));
   auto shape = boost::yap::transform(
       expr, boost::numeric::ublas::detail::transforms::get_extents{});
   if (shape != lhs.extents() && !shape.is_free_scalar()) {
@@ -210,7 +175,8 @@ auto operator/=(boost::numeric::ublas::tensor<T, V, F> &lhs, Expr &&e) {
 }
 
 template <boost::yap::expr_kind K, typename Tuple>
-bool operator!(boost::numeric::ublas::detail::tensor_expression<K, Tuple>&& expr){
+bool operator!(
+    boost::numeric::ublas::detail::tensor_expression<K, Tuple> &&expr) {
   bool result = expr;
   return !result;
 }
