+++ include/boost/numeric/ublas/tensor/multi_index_utility.hpp
@@ -15,7 +15,7 @@
 
 #include <tuple>
 #include <type_traits>
-
+#include <vector>
 
 namespace boost   {
 namespace numeric {
