+++ test/tensor/Jamfile
@@ -29,8 +29,8 @@ test-suite boost-ublas-tensor-test
           test_strides.cpp
           test_expression_comparision1.cpp
           test_expression_comparision2.cpp
-          test_expression_operator.cpp
-          test_expression_operator_decimals.cpp
+          test_expression_operator1.cpp
+          test_expression_operator2.cpp
           test_multiplication.cpp
           test_multi_index_utility.cpp
           test_multi_index.cpp
